requirements.txt
asgiref==3.2.7
boto3==1.12.34
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
jmespath==0.9.5
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
